implement crud and logic insert to table ledge room database

+++ app/src/main/java/com/babsnet/accounting/repository/JournalRepository.kt
@@ -2,12 +2,46 @@ package com.babsnet.accounting.repository
 
 import com.babsnet.accounting.data.entity.Journal
 import com.babsnet.accounting.data.dao.JournalDao
+import com.babsnet.accounting.data.dao.LedgerDao
+import com.babsnet.accounting.data.entity.Account
+import com.babsnet.accounting.data.entity.JournalWithDetails
+import com.babsnet.accounting.data.entity.Ledger
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.runBlocking
+import java.util.Date
 
-class JournalRepository(private val journalDao: JournalDao) {
+class JournalRepository(
+    private val journalDao: JournalDao,
+    private val ledgerDao: LedgerDao
+)
+{
 
-    // gatAllData
-    val allJournals: Flow<List<Journal>> = journalDao.getAllJournals()
+
+    fun getAllJournalsWithDetails(): Flow<List<JournalWithDetails>> {
+        return journalDao.getAllJournals().map { journals ->
+            journals.map { journal ->
+                val ledgers = runBlocking {
+                    journalDao.getLedgersWithAccountByJournalId(journal.journalId)
+                }
+                JournalWithDetails(journal, ledgers)
+            }
+        }
+    }
+
+    fun getJournalByIdWithDetails(journalId: Int): Flow<JournalWithDetails?> {
+        return journalDao.getJournalById(journalId).map { journal ->
+            if (journal != null) {
+                val ledgers = runBlocking {
+                    journalDao.getLedgersWithAccountByJournalId(journal.journalId)
+                }
+                JournalWithDetails(journal, ledgers)
+            } else {
+                null
+            }
+        }
+    }
+
 
     // gateDate based id
     fun getJournalById(journalId: Int): Flow<Journal?> {
@@ -24,10 +58,72 @@ class JournalRepository(private val journalDao: JournalDao) {
     }
 
     suspend fun delete(journal: Journal) {
-        journalDao.delete(journal)
+        val deleteLedger : Int = ledgerDao.deleteAllLedgersByJournal(journal.journalId)
+        if (deleteLedger != 0 ){
+            journalDao.delete(journal)
+        }
     }
 
     suspend fun deleteAll() {
         journalDao.deleteAll()
+    }
+
+    suspend fun saveJournalWithLedger(
+        journal: Journal,
+        debit: Double,
+        credit: Double,
+        account: Account
+    ): Journal? {
+        val ledgers = mutableListOf<Ledger>()
+        val ledger = Ledger(
+            journalId = 0,
+            accountId = account.accountId,
+            debit = debit,
+            credit = credit,
+            createdAt = Date(),
+            updatedAt = null,
+            createdBy = "User",
+            updatedBy = null
+        )
+
+        ledgers.add(ledger)
+
+        return journalDao.insertJournalWithLedger(journal, ledgers)
+    }
+
+
+    suspend fun updateJournalWithLedger(
+        journal: Journal,
+        accountIds: List<Int?>,
+        debit: Double,
+        credit: Double
+    ) {
+
+        if (accountIds.isEmpty()) {
+            throw IllegalArgumentException("Account IDs is empty")
+        }
+
+
+    }
+
+    suspend fun updateJournalLedger(
+        existJournal: Journal,
+        debit: Double,
+        credit: Double,
+        account: Account,
+        existAccountId: Int?
+    ): Journal {
+        journalDao.update(existJournal)
+        val existingLedgers = journalDao.getLedgersByJournalId(existJournal.journalId)
+        if(existingLedgers.isNotEmpty()) {
+            val existingLedger = ledgerDao.getLedgerByJournalIdAndAccountId(existJournal.journalId, existAccountId!!)
+            existingLedger.accountId = account.accountId
+            existingLedger.updatedBy = "User"
+            existingLedger.debit = debit
+            existingLedger.credit = credit
+            existingLedger.updatedAt = Date()
+            ledgerDao.updateLedger(existingLedger)
+        }
+        return existJournal
     }
 }
